+++ src/main/java/kr/co/fastcampus/fastcatch/common/security/CustomUserDetails.java
@@ -1,4 +1,4 @@
-package kr.co.fastcampus.fastcatch.common.config.jwt;
+package kr.co.fastcampus.fastcatch.common.security;
 
 import java.util.Collection;
 import java.util.Collections;
