+++ src/main/java/kr/co/fastcampus/fastcatch/domain/cart/controller/CartController.java
@@ -1,7 +1,7 @@
 package kr.co.fastcampus.fastcatch.domain.cart.controller;
 
 import jakarta.validation.Valid;
-import kr.co.fastcampus.fastcatch.common.config.jwt.CustomUserDetails;
+import kr.co.fastcampus.fastcatch.common.security.CustomUserDetails;
 import kr.co.fastcampus.fastcatch.common.response.ResponseBody;
 import kr.co.fastcampus.fastcatch.domain.cart.dto.request.CartItemRequest;
 import kr.co.fastcampus.fastcatch.domain.cart.dto.response.CartResponse;
@@ -14,7 +14,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
